requirements.txt
discord.py
pandas
flask
jholiday